requirements.txt
aiohttp==3.9.1
aiohttp-retry==2.8.3
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.2.0
certifi==2023.11.17
charset-normalizer==3.3.2
discord==2.3.2
discord.py==2.3.2
frozenlist==1.4.1
idna==3.6
multidict==6.0.4
praw==7.7.1
prawcore==2.4.0
PyJWT==2.8.0
python-dotenv==1.0.1
requests==2.31.0
update-checker==0.18.0
urllib3==2.1.0
websocket-client==1.7.0
yarl==1.9.4
